Update README.md
This code is to understand SAPUI5. SAPUI5 is a set of libraries to build responsive web applications that run on multiple devices like Desktop, Mobile, and Tablet. SAPUI5 Applications follow Model View Controller (MVC) Architecture. In MVC, from UI5 perspective, View and Controller are the Views (XML, HTML, JavaScript and JSON Views) and Controllers (JavaScript Controllers) respectively of the UI5 Application. In PRSES-Application, role of Model is played by OData.